feat(Invoice Maker): add customer management functionality to the app
- Add new files for Customer model and its details
- Add new files for Customer views (CustomersView and CustomerFormView)
- Add new files for Customer sample data
- Update existing files to include Customer model and its views
- Update Product model and its details to change price type from Double to Int
- Update ContentView to include CustomersView
- Update ProductsView to fix a bug in presenting ProductFormView
- Update SettingsBusinessDetailsView to increase line limit for address field

This commit is necessary to provide the functionality of managing customers in the app. It allows the user to add, view, and update customer details. The change in the Product model's price type is to align with the actual use case where price is usually an integer.

+++ Invoice Maker/Model/Product/Product.swift
@@ -13,10 +13,10 @@ class Product {
     @Attribute(.unique) var code: Int
     var name: String
     var details: String
-    var price: Double
+    var price: Int
     var createdDate: Date = Date.now
 
-    init(code: Int, name: String, details: String, price: Double) {
+    init(code: Int, name: String, details: String, price: Int) {
         self.code = code
         self.name = name
         self.details = details
@@ -26,7 +26,10 @@ class Product {
     convenience init?(from productDetails: ProductDetails) {
         if let code = productDetails.code,
            let price = productDetails.price {
-            self.init(code: code, name: productDetails.name, details: productDetails.details, price: Double(price))
+            self.init(code: code,
+                      name: productDetails.name,
+                      details: productDetails.details,
+                      price: price)
         } else {
             return nil
         }
@@ -38,7 +41,7 @@ extension Product {
         if let code = productDetails.code,
            let price = productDetails.price {
             self.code = code
-            self.price = Double(price)
+            self.price = price
         }
 
         name = productDetails.name
